support float values for hr_values() return
supports data from Fossil watches exported through Google Fit

+++ tcxparser/tcxparser.py
@@ -18,7 +18,7 @@ class TCXParser:
 
     def hr_values(self):
         return [
-            int(x.text)
+            float(x.text)
             for x in self.root.xpath(
                 "//ns:HeartRateBpm/ns:Value", namespaces={"ns": namespace}
             )
